Turn vertex slices into Python slices.
Converts the `Slice` object so that it subclasses from `namedtuple`.  This
allows precomputation of `n_atoms` and the creation of a Python `slice`.

The intended use of the additional slice object is something like this:

```python
def do_something_with_slice(vertex_slice):
    # ...
    prepared_data[vertex_slice.as_slice]
    # ...
```

Allowing the vertex slice to be used directly to slice a Python array.  Saves
doing this:

```python
def do_something_with_slice(vertex_slice):
    # ...
    prepared_data[vertex_slice.lo_atom:vertex_slice.hi_atom + 1]
    # ...
```