Unexpected hitch fixed.
The XML reading code in Python does something strange when faced with
XML Namespaces, even where those namespaces have the empty local prefix.
Unfortunately, the only way to enable schema validation is to use XML
namespaces, and so this all interacts horribly. This fixes it in an ugly
way, but until https://bugs.python.org/issue18304 is fixed, we can't fix
this in a better way.