Write sphinx config right
The config was only right with some versions of Sphinx; with the version
deployed on readthedocs, it blew up (trying to split a boolean by
commas, which doesn't work).

None of this is at all helped by the fact that the configuration
settings aren't very well described for programmatic use; there's a lot
of outdated and plain bad descriptions out there.